implement native iterator
If keyAsBuffer or valueAsBuffer, try to coerce most types into a
String. Drop redundant support for the raw option. Just pass
keyAsBuffer = false and/or valueAsBuffer = false.

Make sure that each callback passed to _next() is called exactly
once, and always in the next tick.

Support reopening the indexedDB cursor if it timed out before the
end of the range was reached via reopenOnTimeout option. A snapshot
is created once per cursor so default to false.